evict/requirements.txt
jishaku 
yt-dlp
timezonefinder
humanfriendly
arrow
git+https://github.com/resentbot/uwuipy.git
git+https://github.com/cop-discord/rivalapi.git
git+https://github.com/resentbot/resent.api.git
git+https://github.com/resentbot/discord.py.git
git+https://github.com/resentbot/twscrape.git
git+https://github.com/sheslitforme/Discord-iOS
git+https://github.com/prettylittlelies/discord_paginator@227fd915dd19d09422d80bd3aba601ea486d89f5
python-dateutil
requests
aiogtts
aiohttp
geopy
asyncio
random 
json
os
typing
timed
asyncpg
random 
string
io
collections
deep-translator
emoji
re 
itertools
dataclasses
traceback
math
textwrap
pomice
contextlib
sys
python-dotenv